+++ webcomponents/markdown/CHANGELOG.md
@@ -1,3 +1,14 @@
+# 1.1.0 (2020-11-16)
+
+### Features
+
+- add a method to expose the container
+
+### Style
+
+- improve font styling
+- preserve new lines if `editable`
+
 # 1.0.0 (2020-11-12)
 
 Hello World 👋